refactor: Rename selector.rs to selection_params.rs
The module no longer holds a `Selector`; it now holds `SelectionParams`
and its supporting config/context/error types.
